// enums/users/enumRols.ts
export enum EnumRols {
	Administrador = 'admin',
	CommerTeam = 'commerce_team',
	Commerce = 'commerce',
}

export enum EnumCategoryType {
	Principal = 'principal',
	Canal = 'canal',
	Subcanal = 'subcanal',
	Externo = 'externo',
	ComercialI = 'comercial i',
	ComercialII = 'comercial ii',
	ComercialIII = 'comercial iii',
	ComercialVI = 'comercial vi',
	JefeDeEquipoI = 'jefe de equipo i',
	JefeDeEquipoII = 'jefe de equipo ii',
	JefeDeEquipoVI = 'jefe de equipo vi',
}